+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Do not set CustomElement contents in the constructor
 * Make the flow element work when in a scrollable container
 * Wait until `fv-list` has loaded before adding nodes
+* Finally get scaling working
+* Add translation (panning) support too
 
 ## 0.1.0